// src/app/interfaces/common/shop.interface.ts
export interface Shop {
  select: boolean;
  _id?: string;
  name?: string;
  phoneNo?: string;
  email?: string;
  address?: string;
  websiteName?: string;
  slug?: string;
  domain?: string;
  images?: any;
  subDomain?: string;
  category?: string;
  dateString?: string;
  subCategory?: string;
  package?: any;
  theme?: any;
  themeColor?: string;
  owner?: string;
  buildStatus?: string;
  port?: number;
  users?: any[];
}